Move resolveDisableNotifications into config.go
